units added, new way of controlling bullets
one can now add units, who are labeled with the color blue, and bullets,
which are labeled with the color green. units produce bullets, so
bullets can not exist statically.

now you do not controll position of unit of bullet, but the velocity. this ensures that bullets or units
wont make jumps from one position to another and that bullets at least
always start at their origin relative to their unit.

change of velocity is now again controlled by mathematical expressions.
functions are not quite usable, because engine can not controll how
functions affect the bulletmap.

new variables px, py, pz are added to the change velocity expressions.
these represent the direction to the player, so a bullet can chase a
player (sample is in test.bm.json)